feat: Restart project with modern Python and Poetry
This commit marks a major overhaul of the metro project, restarting it as a modern Python application.

The key changes include:

- **Poetry-based Project:** The project is now managed by Poetry, with all dependencies and project settings defined in `pyproject.toml`. The old Gradle and Java files have been removed.
- **Python 3 Conversion:** The existing Python 2 code has been ported to Python 3, with modernizations and bug fixes.
- **Command-Line Interface:** A new CLI has been implemented using Typer, providing a clean and extensible way to interact with the model.
- **`generate` Command:** A `generate` subcommand has been created to produce an initial city model. It takes population and a random seed as input and saves the city's state to a `city.json` file.
- **`render` Command (Placeholder):** A placeholder for a `render` subcommand has been added. It currently reads the `city.json` file and will be extended to generate SVG maps in the future.
- **Versioning via Git:** The `city.json` file serves as the single source of truth for the city model, allowing its history to be versioned using Git. A `README.md` has been added to explain this workflow.
