+++ metro/city.py
@@ -3,8 +3,7 @@ import math
 
 class City:
 
-    def __init__(self, w=10.0, l=10.0, p = 100000):
-        self.width = width
+    def __init__(self, w=10.0, l=10.0, p=100000):
+        self.width = w
         self.length = l
         self.population = p
-
