+++ metro/commands/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the commands directory a Python package.
