--- population/bin/population_test.py
@@ -1,5 +1,0 @@
-import population
-
-m = population.PopulationModel()
-print m
-
