+++ pyproject.toml
@@ -1,0 +1,16 @@
+[tool.poetry]
+name = "metro"
+version = "0.1.0"
+description = ""
+authors = ["Jules <agent@example.com>"]
+
+[tool.poetry.dependencies]
+python = ">=3.8"
+typer = {extras = ["all"], version = "^0.9.0"}
+
+[tool.poetry.scripts]
+metro = "metro.main:app"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
