--- src/main/java/org/rl337/metro/App.java
@@ -1,36 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package org.rl337.metro;
-
-
-import org.rl337.metro.model.Point2D;
-import org.rl337.metro.model.Shape;
-import org.rl337.metro.model.shapes.Circle;
-import org.rl337.metro.model.shapes.Rectangle;
-
-import java.awt.*;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-
-public class App {
-
-    public static void main(String[] args) {
-
-        Renderer renderer = new Renderer(1024, 768);
-
-        ArrayList<Shape> shapeList = new ArrayList<>();
-        for (int x = 0; x < 4; x++) {
-            for (int y = 0; y < 4; y++) {
-
-                shapeList.add(new Circle(new Point2D(10 + 100 * x, 10 + 100 * y), 5));
-            }
-        }
-
-        renderer.render(shapeList, Color.RED);
-        renderer.show();
-    }
-
-}
