--- src/main/java/org/rl337/metro/model/Point2D.java
@@ -1,29 +1,0 @@
-package org.rl337.metro.model;
-
-public class Point2D {
-    public static final Point2D Origin = new Point2D(0, 0);
-
-    private double x;
-    private double y;
-
-    public Point2D(double x, double y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public double getX() {
-        return x;
-    }
-
-    public void setX(double x) {
-        this.x = x;
-    }
-
-    public double getY() {
-        return y;
-    }
-
-    public void setY(double y) {
-        this.y = y;
-    }
-}
